api: Implement orphan sessions
When a session is renewed, a new session is created and the old session is
now marked orphan. Orphan sessions are like normal sessions but they expire after
a certain amount of time (currently 5s) and they can't be renewed. Orphan
sessions solve a problem with multiple browser tabs renewing sessions
simultaneously: If an XHR API call is made with a specific session token and
another tab renewes the session before the API call has completed, the session
token used in the API call is outdated. With orphan sessions even outdated session
tokens can be used for a specific time.